requirements.txt
boto3
python-editor
termcolor